refactor: custom hook
Moved useAvailableCustomProperties out into their own custom hook
